Split web intent LLM routing helpers
